+++ client/components/LoginForm.tsx
@@ -2,32 +2,49 @@ import { Dispatch, SetStateAction, useState } from "react";
 import { useForm } from "react-hook-form";
 import { ErrorMessage } from "@hookform/error-message";
 import Link from "next/link";
-import SubmitBtn from "./SubmitBtn";
-import useLogin from "../hooks/auth/useLogin";
+import postLogin from "../apis/auth/postLogin";
 import { useSetRecoilState } from "recoil";
 import { isLoginState } from "../recoil/authAtom";
 import { useMutation } from "react-query";
-import postLogin from "../apis/auth/postLogin";
-
+import { useRouter } from "next/router";
+import EmailCheckNumberLayout from "./EmailCheckNumberLayout";
 // import { client } from "../client/client";
 
 interface FormValues {
   id: string;
   password: string;
   setFailedMsg: Dispatch<SetStateAction<string>>;
+  // setFailedMsg: string;
 }
 
-export default () => {
+const LoginForm = () => {
   const [idValue, setIdValue] = useState<string>("");
   const [failedMsg, setFailedMsg] = useState<string>("");
+
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm<FormValues>();
 
-  const onValid = ({ id, password, setFailedMsg }: FormValues) => {
-    useLogin({ id, password, setFailedMsg });
+  const router = useRouter();
+  const loginMutation = useMutation(postLogin);
+  const setIsLogin = useSetRecoilState(isLoginState);
+
+  const useLogin = ({ id, password }: FormValues) => {
+    return loginMutation
+      .mutateAsync({ id, password })
+      .then(res => {
+        if (res.token) {
+          localStorage.setItem("token", res.token);
+          setIsLogin(true);
+          setFailedMsg("");
+          router.push("/signup");
+        }
+      })
+      .catch(() => {
+        setFailedMsg("ID 혹은 비밀번호가 일치하지 않습니다.");
+      });
   };
 
   const handleIdChange = (e: any) => {
@@ -46,7 +63,7 @@ export default () => {
 
         <form
           className="relative flex flex-col mt-5 text-base text-gray-500 font-SCDream5"
-          onSubmit={handleSubmit(onValid)}
+          onSubmit={handleSubmit(useLogin)}
         >
           {/* 아이디 입력 */}
           <label htmlFor="id">ID</label>
@@ -72,7 +89,7 @@ export default () => {
             errors={errors}
             name="id"
             render={({ message }) => (
-              <div className="flex flex-row items-end justify-end w-full mt-1 text-xs text-mainOrange h-fit font-SCDream2">
+              <div className="flex flex-row items-end justify-end w-full mt-1 text-xs text-nagativeMessage h-fit font-SCDream2">
                 {message}
               </div>
             )}
@@ -111,7 +128,7 @@ export default () => {
             errors={errors}
             name="password"
             render={({ message }) => (
-              <div className="flex flex-row items-end justify-end w-full mt-1 text-xs text-mainOrange h-fit font-SCDream2">
+              <div className="flex flex-row items-end justify-end w-full mt-1 text-xs text-nagativeMessage h-fit font-SCDream2">
                 {message}
               </div>
             )}
@@ -135,7 +152,14 @@ export default () => {
         </form>
 
         {/* 로그인 실패 메세지 */}
+        {failedMsg.length !== 0 ? (
+          <div className="self-center my-8 text-xs md:text-sm lg:my-5 font-SCDream4 md:mt-12 text-nagativeMessage">
+            {failedMsg}
+          </div>
+        ) : null}
       </div>
     </>
   );
 };
+
+export default LoginForm;
