+++ client/pages/login.tsx
@@ -2,15 +2,14 @@ import LoginForm from "../components/LoginForm";
 import CertifyPageLayout from "../components/CertifyPageLayout";
 import GoogleBtn from "../components/GoogleBtn";
 import KakaoBtn from "../components/KakaoBtn";
-import SubmitBtn from "../components/SubmitBtn";
 
 const Login = () => {
   return (
     <>
       <CertifyPageLayout>
-        <div className="flex flex-col h-[75%] md:h-[85%] lg:h-full justify-between">
+        <div className="flex flex-col justify-between h-fit">
           <LoginForm />
-          <div className="flex flex-col md:items-center">
+          <div className="flex flex-col mt-5 md:mt-12 lg:mt-6 md:items-center">
             <GoogleBtn />
             <KakaoBtn />
           </div>
